backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
lyricsgenius==3.0.1
beautifulsoup4==4.12.2
requests==2.32.5

